requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.10
litellm==1.52.0
anthropic==0.40.0
# faster-whisper==1.0.3  # Requires ctranslate2 — not available on ARM/musllinux; install manually if needed
APScheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
aioresponses==0.7.6
